# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/vocabulary": {
      "get": {
        "summary": "Vocabulary",
        "operationId": "vocabulary",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/guess": {
      "post": {
        "summary": "Guess",
        "operationId": "guess",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/similar": {
      "get": {
        "summary": "Similar",
        "operationId": "similar",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarWordsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GuessResponse": {
        "properties": {
          "word": {
            "type": "string",
            "title": "Word",
            "description": "The word guessed by the player"
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "description": "The score for the guess"
          },
          "found": {
            "type": "boolean",
            "title": "Found",
            "description": "Whether the player has guessed the daily word"
          }
        },
        "type": "object",
        "required": [
          "word",
          "score",
          "found"
        ],
        "title": "GuessResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SimilarWord": {
        "properties": {
          "word": {
            "type": "string",
            "title": "Word",
            "description": "A word similar to the query"
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "description": "The score for the similar word"
          }
        },
        "type": "object",
        "required": [
          "word",
          "score"
        ],
        "title": "SimilarWord"
      },
      "SimilarWordsResponse": {
        "properties": {
          "words": {
            "items": {
              "$ref": "#/components/schemas/SimilarWord"
            },
            "type": "array",
            "title": "Words",
            "description": "Top similar words ordered by rank descending"
          }
        },
        "type": "object",
        "required": [
          "words"
        ],
        "title": "SimilarWordsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VocabularyResponse": {
        "properties": {
          "words": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Words",
            "description": "Words eligible to be the daily word"
          }
        },
        "type": "object",
        "required": [
          "words"
        ],
        "title": "VocabularyResponse"
      }
    }
  }
}